Renamed/Moved AssetsDatasource.js to /scripts/assets gmo-campaign-list.js : Added logic to only allow blockConfig to be set on initial call to function decorate(block ... otherwise the values from blockConfig are overwritten when paginating to next page.
gmo-campaign-details.js : Added campaign Image
gmo-campaign-details.css : Updated CSS to display campaign Image
