+++ blocks/gmo-campaign-details/gmo-campaign-details.css
@@ -86,14 +86,22 @@ body {
         margin-left: 10px;
     }
 }
+
 .campaign-img {
     width: 80px;
     height: 80px;
-    color: darkgray;
-    border: 1px solid black;
-    border-radius: 20px;
-
+    margin-right: 14px;
+    background: #e2e2e2;
+    border-radius: 18px;
 }
+
+.campaign-img img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 18px;
+}
+
 .tab-wrapper {
     display: flex;
     justify-content: space-between;
